Cleanup remote.rs
